feat(fitness): real FitnessCommandHandler; distribution gate green (#68, #69)
fitness-functions registers its validate command through the CL01 contract; aps-cli delegates like topology. The Phase A poka-yoke gate now passes with 0 errors: red commit 3d280bd to green, proving the validator catches the failure it was built for. Experiment bumped to 0.2.0.
