+++ crates/aps-cli/Cargo.toml
@@ -15,7 +15,7 @@ path = "src/main.rs"
 apss-core.workspace = true
 aps-v1-0000-meta = { version = "1.3.0", path = "../../standards/v1/APS-V1-0000-meta" }
 code-topology = { package = "apss-v1-0001-code-topology", version = "0.2.0", path = "../../standards/v1/APS-V1-0001-code-topology" }
-fitness-functions = { package = "apss-v1-0003-fitness-functions", version = "0.1.0", path = "../../standards-experimental/v1/EXP-V1-0003-fitness-functions" }
+fitness-functions = { package = "apss-v1-0003-fitness-functions", version = "0.2.0", path = "../../standards-experimental/v1/EXP-V1-0003-fitness-functions" }
 apss-project-config = { version = "1.0.0", path = "../../standards/v1/APS-V1-0000-meta/substandards/CF01-project-config" }
 apss-distribution = { version = "1.0.0", path = "../../standards/v1/APS-V1-0000-meta/substandards/DI01-distribution" }
 clap.workspace = true
