+++ standards-experimental/v1/EXP-V1-0003-fitness-functions/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "apss-v1-0003-fitness-functions"
-version = "0.1.0"
+version = "0.2.0"
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
