+++ standards-experimental/v1/EXP-V1-0003-fitness-functions/experiment.toml
@@ -4,7 +4,7 @@ schema = "aps.experiment/v1"
 id = "EXP-V1-0003"
 name = "Architecture Fitness Functions"
 slug = "fitness-functions"
-version = "0.1.0"
+version = "0.2.0"
 category = "technical"
 backwards_compat = true  # Set to false when introducing breaking changes
